Card Overlay
#### Summary
Encapsulate card ```<img>``` in a ```<div>``` and added another
```<div>``` over that to
be able to add text, outlines, or other visual changes.

The overlays are toggled on/off based on the existing suggestion
button.

#### Highlights
* Included ```jQuery.js``` to ```draft.php``` from google CDN to take
advantage of ```toggle()``` function to easily show/hide elements and
expand programmability of future code
* Wrapper ```<div>```s, card ```<img>```, overlay ```<div>```, and
overlay ```<p>``` each have a ```id=``` now to make them findable with
```getElementById``` for future code
* Wrapper ```<div>```s, card ```<img>```, overlay ```<div>```, and
overlay ```<p>``` tags each have their own CSS ```class=``` to make
changes to the appearance easily and findable with
```getElementsByClass```
* Menu items now have CSS (```.nav li```) to set their ```z-index:
99;``` to be above the card overlays

#### Notes
* Any code that was functionally replaced was commented out and labeled
```//Old```
* All ```JS``` and ```jQuery``` added is commented with explanations of
functionality

#### Disclaimer
My code is my best attempt at solving my idea of card overlays. This is
my first experience with JS and CSS at this level of size and
complexity. While the desired effect is achieved the CSS probably needs
a good looking over to make sure it is the best way at defining the
attributes, although I'm sure I tried every combination of attributes
to get the flow to work correct.(CSS is the work of the devil)